separated rospy / py trick from test.
