+++ src/rostful_node/rosinterface/__init__.py
@@ -1,6 +1,8 @@
 #This python package is handling all ROS related communication for rostful-node.
 from __future__ import absolute_import
 
+# TODO : import ( or copy ) rostest_nose here => SAFE enough ?
+
 from .topic import TopicBack
 from .action import ActionBack
 from .service import ServiceBack
