+++ tests/rostful_node/testRosInterface.py
@@ -1,119 +1,24 @@
 #!/usr/bin/env python
-import unittest
+from __future__ import absolute_import
+
+
 import sys
 import os
-import logging
-import nose
-import collections
-import multiprocessing
-import time
+# to be able to run from source
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'src')))
 
-# trying hard to find our ros code ( different possible run environments )
-# We need to support :
-# - rostest
-# - python -m <package> file
-# - nosetests
-# - pycharm UI test runs (easy debug)
-# - TODO : tox + py.test (benchmark)
+# To simulate ROS env setup
+from rostful_node import rostest_nose
+# ROS imports should now work from ROS or from python (without ROS env setup)
+import rospy
+import roslaunch
+from std_msgs.msg import String, Empty
 
-try:
-    import rosgraph
-    import rospy
-    import rostopic
-    import rostest
-    import roslaunch
-    from std_msgs.msg import String, Empty
-    from rostful_node.rosinterface import RosInterface
+# Unit test import
+from rostful_node.rosinterface import RosInterface
 
-except ImportError, ie:
-    logging.warn("{exc}".format(exc=ie))
+import unittest
 
-    # If we cannot import, we assume ros environment not setup. We attempt a custom setup for tests here.
-
-    # setting up all python paths
-    # CAREFUL : SAME order as setup.bash set the pythonpath
-    pythonpath_roscode = collections.OrderedDict({
-        'install': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'install', 'lib', 'python2.7', 'dist-packages')),
-        'devel': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'devel', 'lib', 'python2.7', 'dist-packages')),
-        'indigo': os.path.abspath('/opt/ros/indigo/lib/python2.7/dist-packages'),
-        'src': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'src')),  # should be last since this comes after all other usual ros settings
-    })
-
-    for k, p in pythonpath_roscode.iteritems():
-        if os.path.exists(p):
-            logging.warn("Appending {key} space to python path".format(key=k))
-            sys.path.append(p)
-            # setting python path needed only to find ros shell commands (rosmaster)
-            if p not in os.environ.get("PYTHONPATH", []):
-                os.environ["PYTHONPATH"] = p + ':' + os.environ.get("PYTHONPATH",'')
-
-    # This is enough to fix the import. However all rOS environment should be set before importing rospy due to some static behaviors
-    # So we do the extra setup here
-
-    # Setting env var like ROS if needed
-    # TODO : find the proper place in ros where this is set and use it instead
-    indigo_distro_envvar = {
-        'ROS_ROOT': '/opt/ros/indigo/share/ros',
-        'ROS_PACKAGE_PATH': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..')) + ':/opt/ros/indigo/share:/opt/ros/indigo/stacks',
-        'ROS_MASTER_URI': 'http://localhost:11311',
-        'ROS_DISTRO': 'indigo',
-        'ROS_ETC_DIR': '/opt/ros/indigo/etc/ros',
-    }
-    for k, v in indigo_distro_envvar.iteritems():
-        if os.environ.get(k, None) is None:
-            os.environ[k] = v
-
-    # setting path to find commands
-    ospath_roscode = collections.OrderedDict({
-        'install': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'install', 'bin')),
-        'devel': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'devel', 'bin')),
-        'indigo': os.path.abspath('/opt/ros/indigo/bin'),
-    })
-    ospath_roscode_reversed = collections.OrderedDict(reversed(list(ospath_roscode.items())))  # because we prepend
-    for k, p in ospath_roscode_reversed.iteritems():
-        if os.path.exists(p) and p not in os.environ.get("PATH", []):
-            logging.warn("Appending {key} space to OS path".format(key=k))
-            os.environ["PATH"] = p + ':' +os.environ.get("PATH", '')
-
-    # setting cmake prefix path - rosout needs this
-    cmakepath_roscode = collections.OrderedDict({
-        'install': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'install')),
-        'devel': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'devel')),
-        'indigo': os.path.abspath('/opt/ros/indigo'),
-    })
-    cmakepath_roscode_reversed = collections.OrderedDict(reversed(list(cmakepath_roscode.items())))  # because we prepend
-    for k, p in cmakepath_roscode_reversed.iteritems():
-        if os.path.exists(p) and p not in os.environ.get("CMAKE_PREFIX_PATH", []):
-            logging.warn("Appending {key} space to CMake prefix path".format(key=k))
-            os.environ["CMAKE_PREFIX_PATH"] = p + ':' + os.environ.get("CMAKE_PREFIX_PATH", '')
-
-    # setting ldlibrary path - rosout needs this
-    ldlibrarypath_roscode = collections.OrderedDict({
-        'install': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'install', 'lib')),
-        'install_arch': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'install', 'lib', 'x86_64-linux-gnu')),  # Ref : /opt/ros/indigo/_setup_util.sh
-        'devel': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'devel', 'lib')),
-        'devel_arch': os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..', 'devel', 'lib', 'x86_64-linux-gnu')),  # Ref : /opt/ros/indigo/_setup_util.sh
-        'indigo': os.path.abspath('/opt/ros/indigo/lib'),
-        'indigo_arch': os.path.abspath('/opt/ros/indigo/lib/x86_64-linux-gnu'),  # Ref : /opt/ros/indigo/_setup_util.sh
-    })
-    ldlibrarypath_roscode_reversed = collections.OrderedDict(reversed(list(ldlibrarypath_roscode.items())))  # because we prepend
-    for k, p in ldlibrarypath_roscode_reversed.iteritems():
-        if os.path.exists(p) and p not in os.environ.get("LD_LIBRARY_PATH", []):
-            logging.warn("Appending {key} space to LD_LIBRARY_PATH".format(key=k))
-            os.environ["LD_LIBRARY_PATH"] = p + ':' + os.environ.get("LD_LIBRARY_PATH", '')
-
-    import rosgraph
-    import rospy
-    import rostopic
-    import rostest
-    import roslaunch
-    from std_msgs.msg import String, Empty
-    from rostful_node.rosinterface import RosInterface
-
-
-rostest_enabled = False  # default for python or nose runs
-
-roscore_process = None  # default : we dont need to run roscore
 # test node process not setup by default (rostest dont need it here)
 empty_srv_process = None
 trigger_srv_process = None
@@ -121,24 +26,14 @@ trigger_srv_process = None
 
 # This should have the same effect as the <name>.test file for rostest. Should be used only by nose ( or other python test tool )
 def setup_module():
-    if not rostest_enabled:
-        if not rosgraph.masterapi.is_online():
-            global roscore_process
-            # Trying to solve this : http://answers.ros.org/question/215600/how-can-i-run-roscore-from-python/
-            def ros_core_launch():
-                roslaunch.main(['roscore', '--core'])  # same as rostest_main implementation
-
-            roscore_process = multiprocessing.Process(target=ros_core_launch)
-            roscore_process.start()
-
-        # Workaround until https://github.com/ros/ros_comm/pull/711 is merged and released
-        time.sleep(2)
+    if not rostest_nose.is_rostest_enabled():
+        rostest_nose.rostest_nose_setup_module()
 
         # Start roslaunch
         launch = roslaunch.scriptapi.ROSLaunch()
         launch.start()
 
-        # start required nodes - needs to matcht he content of *.test files for rostest to match
+        # start required nodes - needs to match the content of *.test files for rostest to match
         global empty_srv_process, trigger_srv_process
         empty_srv_node = roslaunch.core.Node('pyros', 'emptyService.py', name='empty_service')
         trigger_srv_node = roslaunch.core.Node('pyros', 'triggerService.py', name='trigger_service')
@@ -147,16 +42,14 @@ def setup_module():
 
 
 def teardown_module():
-    if not rostest_enabled:
+    if not rostest_nose.is_rostest_enabled():
         # finishing all process are finished
         if empty_srv_process is not None:
             empty_srv_process.stop()
         if trigger_srv_process is not None:
             trigger_srv_process.stop()
-        if roscore_process is not None:
-            roscore_process.terminate()  # make sure everything is stopped
-        # in case we dont shutdown with roscore stop
-        rospy.signal_shutdown('test complete')
+
+        rostest_nose.rostest_nose_teardown_module()
 
 
 class TestRosInterface(unittest.TestCase):
@@ -290,13 +183,5 @@ class TestRosInterface(unittest.TestCase):
 if __name__ == '__main__':
 
     # Note : Tests should be able to run with nosetests, or rostest ( which will launch nosetest here )
+    rostest_nose.rostest_or_nose_main('test_ros_interface', 'test_all', TestRosInterface)
 
-    # Ros arguments will tell us if we started from ros, or from straight python
-    rosargs = [arg for arg in sys.argv if arg.startswith("__")]
-
-    if len(rosargs) > 0 :
-        rostest_enabled = True
-        rostest.rosrun('test_ros_interface', 'test_all', TestRosInterface)
-    else :
-        #import nose
-        nose.runmodule()
